# go-backend/docs/openapi/wave1.yaml
openapi: 3.0.3
info:
  title: POS Bengkel Go API (Wave 1)
  version: 0.1.0
servers:
  - url: http://localhost:8081
paths:
...
  /api/v1/part-sales:
    post:
      summary: Create part sale transaction and reserve stock by status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                customer_id:
                  type: integer
                sale_date:
                  type: string
                  format: date
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      part_id:
                        type: integer
                      quantity:
                        type: integer
                      unit_price:
                        type: integer
                status:
                  type: string
                actor_user_id:
                  type: integer
                  nullable: true
      responses:
        '200':
          description: Part sale created payload
        '422':
          description: Invalid payload or insufficient stock
    get:
      summary: Part sales index list with pagination for dashboard page
      parameters:
        - in: query
          name: page
          schema:
            type: integer
          required: false
        - in: query
          name: search
          schema:
            type: string
          required: false
        - in: query
          name: status
          schema:
            type: string
          required: false
        - in: query
          name: payment_status
          schema:
            type: string
          required: false
        - in: query
          name: customer_id
          schema:
            type: integer
          required: false
      responses:
        '200':
          description: Part sales payload compatible with Laravel Inertia props
  /api/v1/part-sales/warranties:
    get:
      summary: Part sale warranties index with filters, summary cards, and pagination
      parameters:
        - in: query
          name: page
          schema:
            type: integer
          required: false
        - in: query
          name: search
          schema:
            type: string
          required: false
        - in: query
          name: warranty_status
          schema:
            type: string
          required: false
        - in: query
          name: source_type
          schema:
            type: string
          required: false
        - in: query
          name: item_type
          schema:
            type: string
          required: false
        - in: query
          name: customer_id
          schema:
            type: integer
          required: false
        - in: query
          name: vehicle_id
          schema:
            type: integer
          required: false
        - in: query
          name: mechanic_id
          schema:
            type: integer
          required: false
        - in: query
          name: date_from
          schema:
            type: string
            format: date
          required: false
        - in: query
          name: date_to
          schema:
            type: string
            format: date
          required: false
        - in: query
          name: expiring_in_days
          schema:
            type: integer
          required: false
      responses:
        '200':
          description: Warranties payload compatible with Laravel Inertia props
...